backend/pyproject.toml
[project]
name = "wave-audio-animation-api"
version = "0.1.0"
description = "FastAPI backend for Wave Audio Animation - audio processing, rendering, and export"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "httpx>=0.27.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
